Attempting to move to src-layout since there are multiple modules now. Also added commandline script
