+++ splitnc/src/splitnc/__init__.py
@@ -1,0 +1,1 @@
+from .splitnc import *
